move sampler from graph index to contrib.sampling
