+++ src/graph/graph_apis.cc
@@ -457,21 +457,21 @@ void CAPI_NeighborUniformSample(DGLArgs args, DGLRetValue* rv) {
   *rv = ConvertSubgraphToPackedFunc(subgs);
 }
 
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling")
 .set_body(CAPI_NeighborUniformSample<1>);
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling2")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling2")
 .set_body(CAPI_NeighborUniformSample<2>);
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling4")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling4")
 .set_body(CAPI_NeighborUniformSample<4>);
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling8")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling8")
 .set_body(CAPI_NeighborUniformSample<8>);
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling16")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling16")
 .set_body(CAPI_NeighborUniformSample<16>);
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling32")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling32")
 .set_body(CAPI_NeighborUniformSample<32>);
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling64")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling64")
 .set_body(CAPI_NeighborUniformSample<64>);
-DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphUniformSampling128")
+DGL_REGISTER_GLOBAL("sampling._CAPI_DGLGraphUniformSampling128")
 .set_body(CAPI_NeighborUniformSample<128>);
 
 DGL_REGISTER_GLOBAL("graph_index._CAPI_DGLGraphGetAdj")
